// FRONTEND/chatbot-admin-panel/src/services/api/schemas/lLMModelType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */

export type LLMModelType = typeof LLMModelType[keyof typeof LLMModelType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LLMModelType = {
  CHAT_COMPLETION: 'CHAT_COMPLETION',
  TEXT_GENERATION: 'TEXT_GENERATION',
  EMBEDDING: 'EMBEDDING',
} as const;